+++ php/Terminus/Runner.php
@@ -4,37 +4,31 @@ namespace Terminus;
 
 use Terminus;
 use Terminus\Utils;
-use Terminus\Loggers\Logger;
 use Terminus\Exceptions\TerminusException;
 
 class Runner {
-  public $config;
-  public $extra_config;
-
   private $arguments;
   private $assoc_args;
-  private $colorize;
-
-  /**
-   * @var LoggerInterface
-   */
+  private $config;
+  private $configurator;
   private $logger;
-
-  /**
-   * @var OutputterInterface
-   */
-  private $outputter;
+  private $terminus;
 
   /**
    * Constructs object. Initializes config, colorizaiton, loger, and outputter
    *
+   * @param [array] $config Extra settings for the config property
    * @return [Runner] $this
    */
-  public function __construct() {
-    $this->initConfig();
-    $this->initColorizaiton();
-    $this->initLogger();
-    $this->initOutputter();
+  public function __construct($config = array()) {
+    $this->setConfigurator();
+    $this->setConfig($config);
+    $params          = array(
+      'runner' => $this,
+    );
+    $params          = array_merge($this->config, $params);
+    $this->terminus  = new Terminus($params);
+    $this->logger    = Terminus::getLogger();
   }
 
   /**
@@ -82,12 +76,12 @@ class Runner {
   }
 
   /**
-   * Determines if output is to be colorized
+   * Retrieves the configurator property
    *
-   * @return [boolean] $this->colorize
+   * @return [Configurator] $this->configurator
    */
-  public function inColor() {
-    return $this->colorize;
+  public function getConfigurator() {
+    return $this->configurator;
   }
 
   /**
@@ -146,13 +140,7 @@ class Runner {
       list($command, $final_args, $cmd_path) = $this->findCommandToRun($args);
       $name = implode(' ', $cmd_path);
 
-      if (isset($this->extra_config[$name])) {
-        $extra_args = $this->extra_config[$name];
-      } else {
-        $extra_args = array();
-      }
-
-      $command->invoke($final_args, $assoc_args, $extra_args);
+      $command->invoke($final_args, $assoc_args);
 
     } catch (\Exception $e) {
       if (method_exists($e, 'getReplacements')) {
@@ -174,75 +162,44 @@ class Runner {
   }
 
   /**
-   * Initializes colorization and saves to Runner property
-   *
-   * @return [void]
-   */
-  private function initColorizaiton() {
-    if ($this->config['colorize'] == 'auto') {
-      $this->colorize = !\cli\Shell::isPiped();
-    } else {
-      $this->colorize = $this->config['colorize'];
-    }
-  }
-
-  /**
    * Initializes configurator, saves config data to it
    *
+   * @param [array] $config Config options to set explicitly
    * @return [void]
    */
-  private function initConfig() {
+  private function setConfig($config = array()) {
     $args = array('terminus', '--debug');
     if (isset($GLOBALS['argv'])) {
       $args = $GLOBALS['argv'];
     }
-    $configurator = Terminus::getConfigurator();
 
     // Runtime config and args
-    list($args, $assoc_args, $runtime_config) = $configurator->parseArgs(
+    list($args, $assoc_args, $runtime_config) = $this->configurator->parseArgs(
       array_slice($args, 1)
     );
 
     $this->arguments  = $args;
     $this->assoc_args = $assoc_args;
 
-    $configurator->mergeArray($runtime_config);
+    $this->configurator->mergeArray($runtime_config);
 
-    list($this->config, $this->extra_config) = $configurator->toArray();
+    $this->config = array_merge($this->configurator->toArray(), $config);
   }
 
   /**
-   * Initializes logger and saves it to Terminus property
+   * Sets the configurator property
    *
+   * @param [Configurator] $configurator Configurator object to set
    * @return [void]
    */
-  private function initLogger() {
-    $this->logger = new Logger(array('config' => $this->config));
-    Terminus::setLogger($this->logger);
-  }
-
-  /**
-   * Initializes outputter and saves it to Terminus property
-   *
-   * @return [void]
-   */
-  private function initOutputter() {
-    // Pick an output formatter
-    if ($this->config['format'] == 'json') {
-      $formatter = new Terminus\Outputters\JSONFormatter();
-    } elseif ($this->config['format'] == 'bash') {
-      $formatter = new Terminus\Outputters\BashFormatter();
+  private function setConfigurator($configurator = null) {
+    if (is_null($configurator)) {
+      $this->configurator = new Configurator(
+        TERMINUS_ROOT . '/php/config-spec.php'
+      );
     } else {
-      $formatter = new Terminus\Outputters\PrettyFormatter();
+      $this->configurator = $configurator;
     }
-
-    // Create an output service.
-    $this->outputter = new Terminus\Outputters\Outputter(
-      new Terminus\Outputters\StreamWriter('php://stdout'),
-      $formatter
-    );
-
-    Terminus::setOutputter($this->outputter);
   }
 
 }