+++ php/config-spec.php
@@ -1,18 +1,6 @@
 <?php
 
 return array(
-  'interactive' => array(
-    'runtime' => '',
-    'file'    => '<bool>',
-    'default' => false,
-    'desc'    => 'Gather input interactively',
-  ),
-  'verbose' => array(
-    'runtime' => '',
-    'file'    => '<bool>',
-    'default' => false,
-    'desc'    => 'Show verbose output',
-  ),
   'colorize' => array(
     'runtime' => '',
     'file'    => '<bool>',